% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor_BUGS_building.R
\name{b_TPC}
\alias{b_TPC}
\title{The way NIMBLE does environmental management
makes me want to tear my hair out.}
\usage{
b_TPC(
  data,
  model,
  priors = NULL,
  samplerType = "RW",
  niter = 10000,
  inits = NULL,
  burn = 0,
  constant_list = NULL,
  ...
)
}
\description{
The way NIMBLE does environmental management
makes me want to tear my hair out.
}
