% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_summaries.R
\name{bayesTPC_summary}
\alias{bayesTPC_summary}
\title{Provide thermal performance curve summaries}
\usage{
bayesTPC_summary(
  TPC,
  Temp_interval = NULL,
  summaryOnly = TRUE,
  summaryType = "hdi",
  centralSummary = "median",
  plot = TRUE,
  probs = c(0.05, 0.95),
  burn = 0,
  plotOnly = FALSE,
  traitName = "Trait",
  ...
)
}
\arguments{
\item{TPC}{list, object output from performing MCMC using the \code{bTPC} function.}

\item{Temp_interval}{vector, reference values to use to compute values of the thermal performance curve. If no vector is provided, Temp_interval is set as a sequence from the lowest observed temperature in the data to the highest observed temperature in the data, with 1000 equally spaced points}

\item{summaryOnly}{logical, should the function return only summaries, or the entire matrix of thermal performance curve evaluations AND the summaries? default = TRUE}

\item{summaryType}{character, type of summary used. Currently supported options are "quantile" and "hdi" (default option). Users that wish to use alternatives may consider summaryOnly = FALSE, and using their desired method on the raw TPC evaluations.}

\item{centralSummary}{character, central summary measure used. Currently supported options are "median" (default) and "mean". Users that wish to use alternatives may consider summaryOnly = FALSE, and using their desired method on the raw TPC evaluations.}

\item{plot}{logical, should a plot be created? default = TRUE}

\item{probs}{numeric, represents either the quantiles to be used during summarization (in the form c(lower, upper)) or the credible mass used to compute the highest density intervals}

\item{burn}{numeric, initial number of iterations to be discarded as burn-in (default = 0)}

\item{plotOnly}{logical, should a plot be generated without returning a list of summaries? if plotOnly = TRUE, an invisible NULL is returned and only plot is generated. Default = FALSE}

\item{traitName}{string, name of Trait to be used as y-axis if a plot is generated. default = "Trait"}

\item{...}{additional parameters to be passed as arguments}
}
\value{
return types vary by argument. For plotOnly = TRUE, invisible NULL is returned. For summaryOnly = TRUE, a list of summaries is returned. For summaryOnly = FALSE, a list is returned with summaries as well as the entire set of thermal performance curve evaluations
}
\description{
Provide thermal performance curve summaries using output from \code{bTPC}'s nimble MCMC
}
\details{
This function returns various summaries of the output of the thermal performance curve model, generarted using MCMC samples from the object returned by \code{bTPC}
}
\examples{
## need data to set up example here. placeholder for now
}
