% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_nimble_models.R
\name{briere_tpc_nimble}
\alias{briere_tpc_nimble}
\title{Evaluate Briere TPC model in Nimble}
\usage{
briere_tpc_nimble(params, Temp, posteriorPredictive)
}
\arguments{
\item{params}{named vector, with entries `q`, `T.max`, `T.min`. If posteriorPredictive = TRUE, `sigma.sq` must also be present.}

\item{Temp}{numeric vector, set of temperature values to evaluate the Briere model at}

\item{posteriorPredictive}{logical, should posterior predictive samples be generated instead of evaluating the deterministic Briere model. default = FALSE.}
}
\value{
numeric vector, where entry `k` represents either the value of the Briere model at `Temp[k]` for the given vector `params` (posteriorPredictive = FALSE), or a sigle draw from the posterior predictive distribution of the Schoolfield-Sharpe model at `Temp[k]` for the given vector `params` (posteriorPredictive = TRUE)
}
\description{
Evaluate Briere model for thermal performance for a given parameter set.
}
\details{
This function returns a numeric vector of evaluations of the Briere thermal performance curve for a reference set `Temp`
}
\examples{
## set params and reference temperature set
param_set = c(T.max = 45, T.min = 20, q = .05)
Temp_ref = seq(from = 5, to = 50, length.out = 1000)
plot(Temp_ref, briere_tpc(params = param_set, Temp = Temp_ref), type = 'l')
}
