% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_spec.R, R/model_spec.R
\name{change_constants.btpc_likelihood}
\alias{change_constants.btpc_likelihood}
\alias{change_constants}
\alias{change_constants.btpc_model}
\title{Change constants of pre-specified model}
\usage{
\method{change_constants}{btpc_likelihood}(x, constants)

change_constants(x, constants)

\method{change_constants}{btpc_model}(x, constants)
}
\arguments{
\item{x}{\code{btpc_model} or \code{btpc_likelihood}, The specification to be changed.}

\item{constants}{named character, The names should correspond to the constants to change, and the values should be the new desired constants.}
}
\value{
Returns the modified model. Does not change the default values of any registered model type.
}
\description{
Intended to be used to change the constants of an already existing model.
}
\details{
\code{bayesTPC} does not verify if constants specified are compatible with NIMBLE's dialect of BUGS.
All available distributions and formatting are provided on the
\href{https://r-nimble.org/html_manual/cha-writing-models.html#subsec:dists-and-functions}{NIMBLE user manual}.
}
