% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_specification.R
\name{change_priors}
\alias{change_priors}
\title{Change priors of pre-specified model}
\usage{
change_priors(model, priors)
}
\arguments{
\item{model}{Object of type \code{btpc_model}. The specification to be changed.}

\item{priors}{named character. The names should correspond to the parameters to change, and the values should be the new desired priors.}
}
\value{
Returns the modified model. Does not change the default values of any registered model type.
}
\description{
Intended to be used to change the priors of an already existing model.
}
