% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_exception_handling.R
\name{check_params}
\alias{check_params}
\title{Checks Prior data format}
\usage{
check_params(model, params)
}
\arguments{
\item{model}{character, the model type}

\item{params}{named vector, contains the actual params from the function call}
}
\value{
list, input params coerced as a vector, if necessary
}
\description{
//This documentation may not be necessary.
Check if data is correctly formatted. For interal use.
}
\details{
Helper function to help error correction.
}
