% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_exception_handling.R
\name{checkData}
\alias{checkData}
\title{Check TPC data format}
\usage{
checkData(data)
}
\arguments{
\item{data}{list, with expected entries "Trait" (corresponding to the trait being modeled by the thermal performance curve) and "Temp" (corresponding to the temperature in Celcius that the trait is being measured at).}
}
\value{
list, with entries "data" (object of class "list", data to be passed to \code{nimble} model) and "N" (integer number of data points, to be passed \code{nimble} model as a constant)
}
\description{
Check if data is correctly formatted
}
\details{
This function returns a list of data and constants to be passed to a \code{nimble} model to perform MCMC
}
\examples{
## generate data
test_data <- list(Temp = rep(c(10, 20, 30, 40), 5), Trait = rgamma(20, 5, rep(c(10, 20, 30, 40), 5)))
checkData(test_data)
}
