% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor_BUGS_building.R
\name{configure_model}
\alias{configure_model}
\title{Create model string}
\usage{
configure_model(model, priors = NULL, verbose = TRUE)
}
\arguments{
\item{model}{character, name of thermal performance curve model. Currently, supported options include "quadratic", "briere", "gaussian", "weibull", "pawar-shsch", "lactin2", "kamykowski","ratkowsky", "stinner", "binomial_glm_lin", "binomial_glm_quad".}

\item{priors}{list, optional input specifying prior distributions for parameters (default = NULL). Elements of the list should correspond to model parameters, and written using nimble logic. For parameters not specified in the list, default priors are used.}

\item{verbose}{logical, optional input. If verbose = TRUE, messages are printed when for priors that deviate from the defaultPriors(model) (see ?defaultPriors for additional information). Default = TRUE}
}
\value{
character, character string specifying the default model formulation to be passed to `nimble`.
}
\description{
Create model string for thermal performance curve model to be passed to nimble
}
\details{
This function returns a character string of the full `nimble` model for a user-specified thermal performance curve and prior distributions
}
\examples{
## Print default model for briere
cat(defaultModel(model = 'briere'))

## Use custom prior for 'q' parameter in quadratic curve
my_prior = list(q = 'q~beta(.5, .5)')
cat(defaultModel(model = 'quadratic', priors = my_prior))
}
