% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultPriors.R
\name{defaultModel}
\alias{defaultModel}
\title{Retrieve default model}
\usage{
defaultModel(model)
}
\arguments{
\item{model}{character, name of thermal performance curve model. Currently, supported options include "quadratic", "briere", "gaussian", "weibull", "pawar-shsch", and "lactin2".}
}
\value{
character, character string specifying the default likelihood model formulation to be passed to `nimble`
}
\description{
Retrieve default model formulation for a given thermal performance curve model
}
\details{
This function returns a character string of the `nimble` likelihood model for a user-specified thermal performance curve
}
\examples{
cat(defaultModel(model = 'briere'))

}
