% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_default_priors.R
\name{defaultPriors}
\alias{defaultPriors}
\title{Retrieve default prior choices}
\usage{
defaultPriors(model)
}
\arguments{
\item{model}{character, name of thermal performance curve model. Currently, supported options include "quadratic", "briere", "gaussian", "weibull", "pawar-shsch", "lactin2", "kamykowski", "ratkowsky", "binomial_glm_lin", "binomial_glm_quad",  and "stinner".}
}
\value{
list, named list detailing the default prior distributions for each parameter of the model
}
\description{
Retrieve default prior choices for a given thermal performance curve model
}
\details{
This function returns the default prior choices for the parameters of the thermal performance curve model that is passed as an input argument
}
\examples{
defaultPriors(model = 'quadratic')

}
