% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectraits_api.R
\name{find_VB_datasets}
\alias{find_VB_datasets}
\title{Searches and retrieves Vectraits datasets}
\usage{
find_VB_datasets(keywords, safety = TRUE)
}
\arguments{
\item{keywords}{character, one or more search terms.}

\item{safety}{logical, should an error be thrown if too many datasets are retreived?. Default is TRUE.}
}
\value{
A list of all datasets matching the search terms provided.
}
\description{
Searches and retrieves Vectraits datasets
}
