% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_nimble_models.R
\name{gauss_tpc_nimble}
\alias{gauss_tpc_nimble}
\title{Evaluate Gaussian TPC model in Nimble}
\usage{
gauss_tpc_nimble(params, Temp, posteriorPredictive)
}
\arguments{
\item{params}{named vector, with entries `a`, `rmax`, `T.opt`. If posteriorPredictive = TRUE, `sigma.sq` must also be present.}

\item{Temp}{numeric vector, set of temperature values to evaluate the Gaussian model at}

\item{posteriorPredictive}{logical, should posterior predictive samples be generated instead of evaluating the deterministic Gaussian model. default = FALSE.}
}
\value{
numeric vector, where entry `k` represents either the value of the Gaussian model at `Temp[k]` for the given vector `params` (posteriorPredictive = FALSE), or a sigle draw from the posterior predictive distribution of the Schoolfield-Sharpe model at `Temp[k]` for the given vector `params` (posteriorPredictive = TRUE)
}
\description{
Evaluate Gaussian model for thermal performance for a given parameter set.
}
\details{
This function returns a numeric vector of evaluations of the Gaussian thermal performance curve for a reference set `Temp`
}
\examples{
## set params and reference temperature set
param_set = c(T.opt = 36, a = 6.5, rmax = 2.75)
Temp_ref = seq(from = 5, to = 50, length.out = 1000)
plot(Temp_ref, gauss_tpc(params = param_set, Temp = Temp_ref), type = 'l')
}
