% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectraits_api.R
\name{get_dataset}
\alias{get_dataset}
\title{Retreive a Dataset from VecTraits}
\usage{
get_dataset(ID = -1)
}
\arguments{
\item{ID}{integer, the dataset to retrieve}
}
\value{
A dataframe containing the dataset in Vectraits corresponding to ID
}
\description{
Retreive a Dataset from VecTraits
}
