% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectraits_api.R
\name{get_datasets}
\alias{get_datasets}
\title{Retrieve multiple datasets from VecTraits}
\usage{
get_datasets(IDS)
}
\arguments{
\item{IDS}{integer, the datasets to retrieve.}
}
\value{
A list of the datasets requested.
}
\description{
Retrieve multiple datasets from VecTraits
}
