% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_default_model_specification}
\alias{get_default_model_specification}
\title{Full Model Specification}
\usage{
get_default_model_specification(model)
}
\arguments{
\item{model}{character or \code{btpc_model}. Use \code{\link[=get_models]{get_models()}} to view all options.}
}
\value{
A \code{btpc_model} object containing the model type, formula, parameters and respective priors,
constants and respective default values if applicable, and variance prior if applicable.
}
\description{
Returns the full details of an implemented model.
}
