% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_get_functions.R
\name{get_default_priors}
\alias{get_default_priors}
\title{Get Default Prior Distributions}
\usage{
get_default_priors(model)
}
\arguments{
\item{model}{A string specifying the model name. Use `get_models()` to view all options.}
}
\description{
TODO
}
