% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_get_functions.R
\name{get_formulas}
\alias{get_formulas}
\title{Get Models and Formulas}
\usage{
get_formulas()
}
\value{
\code{get_formulas()} returns a named list containing all models and their respective formulas as expressions.
}
\description{
Returns all implemented models and their respective formulas.
}
\examples{

get_models()
}
