% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_get_functions.R
\name{get_model_constants}
\alias{get_model_constants}
\title{Get Model Constants}
\usage{
get_model_constants(model)
}
\arguments{
\item{model}{A string specifying the model name. Use `get_models()` to view all options.}
}
\description{
TODO
}
