% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_get_functions.R
\name{get_model_function}
\alias{get_model_function}
\title{Get Model as a Function}
\usage{
get_model_function(model)
}
\arguments{
\item{model}{A string specifying the model name. Use \code{\link[=get_models]{get_models()}} to view all options.}
}
\description{
TODO
}
