% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_get_functions.R
\name{get_model_names}
\alias{get_model_names}
\title{Get Model Names}
\usage{
get_model_names()
}
\value{
\code{get_model_names()} returns a character vector of all implemented models in \code{bayesTPC}.
}
\description{
Returns the names of all implemented models in \code{bayesTPC}.
}
\examples{

get_model_names()
}
