% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_get_functions.R
\name{get_models}
\alias{get_models}
\title{Get Models and Formulas}
\usage{
get_models()
}
\value{
\code{get_models()} returns a data frame with two columns containing the names and formulas for all implemented models in \code{bayesTPC}.
}
\description{
Returns all implemented models and their respective formulas.
}
\examples{

get_models()
}
