% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectraits_api.R
\name{get_VB_dataset}
\alias{get_VB_dataset}
\title{Retreive a Dataset from VecTraits}
\usage{
get_VB_dataset(ID = -1, check_interactive = TRUE)
}
\arguments{
\item{ID}{integer, the dataset to retrieve.}

\item{check_interactive}{logical, should the user be prompted to input an ID if one isn't provided? Default is TRUE}
}
\value{
A dataframe containing the dataset in Vectraits corresponding to ID
}
\description{
Retreive a Dataset from VecTraits
}
