% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_generics.R
\name{hist.btpc_MCMC}
\alias{hist.btpc_MCMC}
\title{Plots Posterior Sample Histogram}
\usage{
\method{hist}{btpc_MCMC}(x, plot = T, burn = 0, ...)
}
\arguments{
\item{x}{\code{btpc_MCMC}, object output from performing MCMC using the \code{bTPC} function.}

\item{plot}{logical, should histograms be plotted? Default is TRUE.}

\item{burn}{numeric, initial number of iterations to be discarded as burn-in. Default is 0.}

\item{...}{parameters passed to \link{hist.default}.}
}
\value{
Returns a named list of the \code{histogram} objects for each model parameter. Returned invisibly unless plot = F.
}
\description{
Plots Posterior Sample Histogram
}
