% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ipairs}
\alias{ipairs}
\title{Wrapper for IDPmisc::ipairs() to create pairs plots of bTPC output}
\usage{
ipairs(
  x,
  burn = 0,
  thin = 1,
  ztransf = function(x) {
     x[x < 1] <- 1
     log2(x)
 },
  ...
)
}
\arguments{
\item{x}{\code{mcmc}, \code{matrix}, or \code{list}. If a list, must contain an element called \code{samples} that is of class \code{mcmc} or \code{matrix}.}

\item{burn}{integer, number of samples to discard as burn-in before creating traceplot. Default is 0.}

\item{thin}{integer, thinning interval used to generate traceplots.  Default is 1.}

\item{ztransf}{Function, used to transform the counts per pixel in \code{\link[IDPmisc:ipairs]{IDPmisc::ipairs()}}.}

\item{...}{additional graphical parameters to be passed as arguments to \code{\link[IDPmisc:ipairs]{IDPmisc::ipairs()}}.
For additional information, try \code{?IDPmisc::ipairs}.}
}
\value{
Returns invisible(NULL) and creates pairs plots for MCMC parameters.
}
\description{
Wrapper for IDPmisc() that can both directly accept samples of object type \code{mcmc} or \code{matrix} as well as an object of class \code{list} with an element \code{samples} that is of class \code{mcmc} or \code{matrix}.
}
\details{
This is a wrapper to create pairs plots using IDPmisc's \code{ipairs()} function.
}
