% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_TPC.R
\name{MAP_estimate}
\alias{MAP_estimate}
\title{Find Maximum A Posteriori Estimate.}
\usage{
MAP_estimate(fit)
}
\arguments{
\item{fit}{\code{btpc_MCMC}, output from \code{\link[=b_TPC]{b_TPC()}}}
}
\value{
\code{MAP_estimate} returns the parameters of the sample with the highest posterior density.
}
\description{
Find Maximum A Posteriori Estimate.
}
