% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultPriors.R
\name{pawar_shsch_tpc}
\alias{pawar_shsch_tpc}
\title{Evaluate Schoolfield-Sharpe model}
\usage{
pawar_shsch_tpc(params, Temp, T.ref, posteriorPredictive = FALSE)
}
\arguments{
\item{params}{named vector, with entries `a`, `b`, `c`, `T.min`, `T.max`. If posteriorPredictive = TRUE, `sigma.sq` must also be present.}

\item{Temp}{numeric vector, set of temperature values to evaluate the Schoolfield-Sharpe model at}

\item{T.ref}{numeric, value for fixed parameter T.ref  used in the formulation of the Schoolfield-Sharpe model from Kontopoulos et al, 2018}

\item{posteriorPredictive}{logical, should posterior predictive samples be generated instead of evaluating the deterministic Schoolfield-Sharpe model. default = FALSE.}
}
\value{
numeric vector, where entry `k` represents either the value of the Schoolfield-Sharpe model at `Temp[k]` for the given vector `params` (posteriorPredictive = FALSE), or a sigle draw from the posterior predictive distribution of the Schoolfield-Sharpe model at `Temp[k]` for the given vector `params` (posteriorPredictive = TRUE)
}
\description{
Evaluate Schoolfield-Sharpe model for thermal performance for a given parameter set using the Schoolfield-Sharpe formulation from Kontopoulos et al, 2018
}
\details{
This function returns a numeric vector of evaluations of the Schoolfield-Sharpe thermal performance curve for a reference set `Temp`
}
\examples{
## set params and reference temperature set
param_set = c(T.opt = 39, e = .7, e_h = 15, r_tref = 1)
Temp_ref = seq(from = 5, to = 50, length.out = 1000)
plot(Temp_ref, pawar_shsch_tpc(params = param_set, Temp = Temp_ref, T.ref = 20), type = 'l')
}
