% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_generics.R
\name{posterior_predictive}
\alias{posterior_predictive}
\alias{plot_prediction.btpc_MCMC}
\title{Provide thermal performance curve posterior predictive summaries}
\usage{
posterior_predictive(
  TPC,
  temp_interval = NULL,
  summaryType = "hdi",
  centralSummary = "median",
  prob = 0.9,
  quantiles = c(0.05, 0.95),
  burn = 0,
  type = "response",
  seed = NULL
)

\method{plot_prediction}{btpc_MCMC}(
  x,
  ylab = "Trait",
  legend = TRUE,
  legend_position = "bottomright",
  temp_interval = NULL,
  summaryType = "hdi",
  centralSummary = "median",
  prob = 0.9,
  quantiles = c(0.05, 0.95),
  burn = 0,
  seed = NULL,
  ...
)
}
\arguments{
\item{TPC}{\code{btpc_MCMC}, object output from performing MCMC using the \code{b_TPC} function.}

\item{temp_interval}{numeric, reference values to use to compute values of the thermal performance curve.
If no vector is provided, temp_interval is set as a sequence from the lowest observed temperature in the data to the highest observed temperature in the data, with 1,000 equally spaced points.}

\item{summaryType}{character, Determines what method is used to create credible intervals. Currently supported options are "quantile" and "hdi" (default).}

\item{centralSummary}{character, central summary measure used. Currently supported options are "median" (default) and "mean".}

\item{prob}{numeric, the credible mass used to compute the highest density interval. Used if summaryType = "hdi".}

\item{quantiles}{numeric, quantiles used for a credible interval. Used if summaryType = "quantile".}

\item{burn}{numeric, initial number of iterations to be discarded as burn-in. Default is 0.}

\item{type}{character, should the summaries be calculated for the link or the response?
Supported inputs are "response" and "link". Default is "response".}

\item{seed}{integer, seed value to be used. Useful for ensuring that results are reproducible. Default is NULL.}

\item{x}{\code{btpc_prediction} or \code{btpc_MCMC}, output from \code{\link[=posterior_predictive]{posterior_predictive()}} or \code{\link[=b_TPC]{b_TPC()}}, respectively.}

\item{ylab}{character, a title for the y-axis. Default is "Trait".}

\item{legend}{logical, should a legend be added to the plot? Default is TRUE.}

\item{legend_position}{character, position of the legend. Only used if legend = TRUE. Default is "bottomright".}

\item{...}{additional parameters passed to \code{plot()}.}
}
\value{
A list containing the central summary and the bounds of the credible interval generated by the posterior samples.
}
\description{
Provide thermal performance curve posterior predictive summaries using output from \code{\link[=b_TPC]{b_TPC()}}'s nimble MCMC.
}
\details{
This function returns various summaries of the output of the thermal performance curve posterior predictive model samples, generated using MCMC samples from the object returned by \code{b_TPC()}.
}
