% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_summaries.R
\name{posteriorPredTPC}
\alias{posteriorPredTPC}
\title{Provide thermal performance curve posterior predictive summaries}
\usage{
posteriorPredTPC(
  TPC,
  Temp_interval = NULL,
  summaryOnly = TRUE,
  summaryType = "hdi",
  centralSummary = "median",
  plot = TRUE,
  probs = c(0.05, 0.95),
  burn = 0,
  plotOnly = FALSE,
  traitName = "Trait",
  seed = NULL,
  ...
)
}
\arguments{
\item{TPC}{list, object output from performing MCMC using the \code{bTPC} function.}

\item{Temp_interval}{vector, reference values to use to compute values of the thermal performance curve.
If no vector is provided, Temp_interval is set as a sequence from the lowest observed temperature in the data to the highest observed temperature in the data, with 1,000 equally spaced points.}

\item{summaryOnly}{logical, should the function return only summaries, or the entire matrix of thermal performance curve posterior predictive samples AND the summaries? Default is TRUE.}

\item{summaryType}{character, type of summary used. Currently supported options are "quantile" and "hdi" (default option).
Users that wish to use alternatives may consider summaryOnly = FALSE, and using their desired method on the raw TPC evaluations.}

\item{centralSummary}{character, central summary measure used. Currently supported options are "median" (default) and "mean".
Users that wish to use alternatives may consider summaryOnly = FALSE, and using their desired method on the raw TPC evaluations.}

\item{plot}{logical, should a plot be created? Default is TRUE.}

\item{probs}{numeric, represents either the quantiles to be used during summarization (in the form c(lower, upper)) or the credible mass used to compute the highest density intervals.}

\item{burn}{numeric, initial number of iterations to be discarded as burn-in. Default is 0.}

\item{plotOnly}{logical, should a plot be generated without returning a list of summaries? if plotOnly = TRUE, an invisible NULL is returned and only plot is generated. Default is FALSE.}

\item{traitName}{string, name of Trait to be used as y-axis if a plot is generated. Default is "Trait".}

\item{seed}{integer, seed value to be used. Useful for ensuring that results are reproducible. Default is NULL.}

\item{...}{additional parameters to be passed as arguments.}
}
\value{
return types vary by argument. For plotOnly = TRUE, invisible NULL is returned. For summaryOnly = TRUE, a list of summaries is returned.
For summaryOnly = FALSE, a list is returned with summaries as well as the entire set of thermal performance curve posterior predictive samples.
}
\description{
Deprecated. Please use \code{\link[=posterior_predictive]{posterior_predictive()}} and \code{\link[=plot_prediction]{plot_prediction()}} instead.
Provide thermal performance curve posterior predictive summaries using output from \code{\link[=b_TPC]{b_TPC()}}'s nimble MCMC.
}
\details{
This function returns various summaries of the output of the thermal performance curve posterior predictive model samples, generated using MCMC samples from the object returned by \code{b_TPC()}.
}
\examples{
## need data to set up example here. placeholder for now
}
