% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultPriors.R
\name{quadratic_tpc}
\alias{quadratic_tpc}
\title{Evaluate Quadratic TPC model}
\usage{
quadratic_tpc(params, Temp, posteriorPredictive = FALSE)
}
\arguments{
\item{params}{named vector, with entries `q`, `T.max`, `T.min`. If posteriorPredictive = TRUE, `sigma.sq` must also be present.}

\item{Temp}{numeric vector, set of temperature values to evaluate the Quadratic model at}

\item{posteriorPredictive}{logical, should posterior predictive samples be generated instead of evaluating the deterministic Quadratic model. default = FALSE.}
}
\value{
numeric vector, where entry `k` represents either the value of the Quadratic model at `Temp[k]` for the given vector `params` (posteriorPredictive = FALSE), or a sigle draw from the posterior predictive distribution of the Schoolfield-Sharpe model at `Temp[k]` for the given vector `params` (posteriorPredictive = TRUE)
}
\description{
Evaluate Quadratic model for thermal performance for a given parameter set.
}
\details{
This function returns a numeric vector of evaluations of the Quadratic thermal performance curve for a reference set `Temp`
}
\examples{
## set params and reference temperature set
param_set = c(T.max = 47, T.min = 23, q = .02)
Temp_ref = seq(from = 5, to = 50, length.out = 1000)
plot(Temp_ref, quadratic_tpc(params = param_set, Temp = Temp_ref), type = 'l')
}
