% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_spec.R
\name{remove_likelihood}
\alias{remove_likelihood}
\alias{reset_likelihoods}
\title{Removing User-Defined Likelihoods}
\usage{
remove_likelihood(name)

reset_likelihoods()
}
\arguments{
\item{name}{character, the name of the likelihood to be removed}
}
\description{
Removes one or all user-defined likelihoods accessible to other \code{bayesTPC} functions. \code{remove_likelihood()} only removes one specific likelihood, while \code{reset_likelihoods()} removes \strong{all} user-defined likelihoods.
}
