% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{remove_model}
\alias{remove_model}
\alias{reset_models}
\title{Removing User-defined models}
\usage{
remove_model(name)

reset_models()
}
\arguments{
\item{name}{character, the name of the model to be removed}
}
\description{
Removes one or all user-defined models accessible to other \code{bayesTPC} functions. \code{remove_model()} only removes one specific model, while \code{reset_models()} removes \strong{all} user-defined models.
}
\seealso{
\code{\link[=specify_model]{specify_model()}}, \code{\link[=specify_bernoulli_model]{specify_bernoulli_model()}}, \code{\link[=specify_normal_model]{specify_normal_model()}}, \code{\link[=specify_binomial_model]{specify_binomial_model()}}
}
