% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{specify_model}
\alias{specify_model}
\title{Specify a model}
\usage{
specify_model(
  name = character(),
  parameters = character(),
  formula = expression(),
  constants = double(),
  link = "identity",
  distribution = "normal",
  llh_parameters = character(),
  llh_constants = double(),
  ...
)
}
\arguments{
\item{name}{character, The desired name of the model specification.}

\item{parameters}{named character, the names should correspond to the parameters being fit,
and the values should be the prior distributions to be drawn from for each respective parameter.
Uniform distributions should be used unless there is good reason to draw from another.}

\item{formula}{expression, The actual formula being fit.
Must include 'Temp' to represent temperature and all specified parameters and constants}

\item{constants}{optional named double, Represents any terms in the formula that should not be fit.}

\item{link}{character, A link function between the model fit and the trait values.
Currently supported options are: 'identity', 'log', 'logit', and 'reciprocal'. Default is 'identity'.}

\item{distribution}{character, The distribution used to calculate likelihoods.
Currently supported options are: 'normal', 'poisson', 'bernoulli', 'binomial', 'exponential', and 'gamma'. Default is 'normal'.}

\item{llh_parameters}{optional named character, additional parameters used in the likelihood calculation (names) and their respective priors (value).}

\item{llh_constants}{optional named double, additional constants used in the likelihood calculation (names) and their respective default values (value).}

\item{...}{Additional model specification attributes.}
}
\value{
Returns an object of type \code{btpc_model}, which can then be used in other \code{bayesTPC} functions.
The model name is also registered, and so can be accessed using by passing only the name into functions.
However, user-defined models are not saved between sessions, and will be reset whenever the package is reloaded.
}
\description{
Creates an object with the required formatting to be fit using other \code{bayesTPC} functions.
}
\details{
\code{bayesTPC} does not verify if the priors specified are compatible with NIMBLE's dialect of BUGS.
All available distributions and formatting are provided on the
\href{https://r-nimble.org/html_manual/cha-writing-models.html#subsec:dists-and-functions}{NIMBLE user manual}.

This model type should be used for counts of binary results.
}
\examples{
my_name <- "my_model"
my_formula <- expression(a * Temp^c + b)
my_parameters <- c(a = "dunif(0,1)", b = "dnorm(0,1)")
my_constants <- c(c = 1.5)

\dontrun{
my_model <- specify_model(
  name = my_name,
  parameters = my_parameters,
  formula = my_formula,
  constants = my_constants,
  link = "logit",
  distribution = "binomial"
)
}
}
\seealso{
\code{\link[=remove_model]{remove_model()}}, \code{\link[=reset_models]{reset_models()}}, \code{\link[=specify_bernoulli_model]{specify_bernoulli_model()}}, \code{\link[=specify_normal_model]{specify_normal_model()}}, \code{\link[=specify_binomial_model]{specify_binomial_model()}}
}
