% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_model_evaluation.R
\name{stinner_tpc}
\alias{stinner_tpc}
\title{Evaluate Stinner model}
\usage{
stinner_tpc(params, Temp, posteriorPredictive = FALSE)
}
\arguments{
\item{params}{named vector, with entries `C`, `k1`, `k2`, `T.opt`. If posteriorPredictive = TRUE, `sigma.sq` must also be present.}

\item{Temp}{numeric vector, set of temperature values to evaluate the Stinner model at}

\item{posteriorPredictive}{logical, should posterior predictive samples be generated instead of evaluating the deterministic Stinner model. default = FALSE.}
}
\value{
numeric vector, where entry `k` represents either the value of the Stinner model at `Temp[k]` for the given vector `params` (posteriorPredictive = FALSE), or a sigle draw from the posterior predictive distribution of the Stinner model at `Temp[k]` for the given vector `params` (posteriorPredictive = TRUE)
}
\description{
Evaluate Stinner model for thermal performance for a given parameter set
}
\details{
This function returns a numeric vector of evaluations of the Stinner thermal performance curve for a reference set `Temp`
}
\examples{
## set params and reference temperature set
param_set = c(T.max = 47, T.min = 10, a = .06, b = .5)
Temp_ref = seq(from = 5, to = 50, length.out = 1000)
plot(Temp_ref, ratkowsky_tpc(params = param_set, Temp = Temp_ref), type = 'l')
}
