% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultPriors.R
\name{str2tpc_fun}
\alias{str2tpc_fun}
\title{Extract function from model string}
\usage{
str2tpc_fun(model)
}
\arguments{
\item{model}{character, name of thermal performance curve model. Currently, supported options include "quadratic", "briere", "gaussian", "weibull", "pawar-shsch", "lactin2", "kamykowski" and "ratkowsky".}
}
\value{
function handle, corresponding to function used to evaluate the thermal performance curve for a given `model` input
}
\description{
Extract thermal performance curve function from model string used to specify type of model to fit
}
\details{
This function returns a function that can be used to evaluate a thermal performance for a model. Designed to be called internally by `posteriorPredTPC()` and `bayesTPC_summary()`. Not recommended to be called directly, though it is possible (see examples)
}
\examples{
## set params and reference temperature set
myfun = str2tpc_fun(model = 'gaussian')
param_set = c(T.opt = 36, a = 6.5, rmax = 2.75)
Temp_ref = seq(from = 5, to = 50, length.out = 1000)
plot(Temp_ref, myfun(params = param_set, Temp = Temp_ref), type = 'l')
}
