% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultPriors.R
\name{weibull_tpc}
\alias{weibull_tpc}
\title{Evaluate Weibull TPC model}
\usage{
weibull_tpc(params, Temp, posteriorPredictive = FALSE)
}
\arguments{
\item{params}{named vector, with entries `a`, `b`, `c`, `T.opt`. If posteriorPredictive = TRUE, `sigma.sq` must also be present.}

\item{Temp}{numeric vector, set of temperature values to evaluate the Weibull model at}

\item{posteriorPredictive}{logical, should posterior predictive samples be generated instead of evaluating the deterministic Weibull model. default = FALSE.}
}
\value{
numeric vector, where entry `k` represents either the value of the Weibull model at `Temp[k]` for the given vector `params` (posteriorPredictive = FALSE), or a sigle draw from the posterior predictive distribution of the Schoolfield-Sharpe model at `Temp[k]` for the given vector `params` (posteriorPredictive = TRUE)
}
\description{
Evaluate Weibull model for thermal performance for a given parameter set.
}
\details{
This function returns a numeric vector of evaluations of the Weibull thermal performance curve for a reference set `Temp`
}
\examples{
## set params and reference temperature set
param_set = c(T.opt = 38, a = 3, b = 7e9, c = 1e9)
Temp_ref = seq(from = 5, to = 50, length.out = 1000)
plot(Temp_ref, weibull_tpc(params = param_set, Temp = Temp_ref), type = 'l')
}
