% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_BUGS_building.R
\name{b_TPC}
\alias{b_TPC}
\title{Perform MCMC}
\usage{
b_TPC(
  data,
  model,
  priors = NULL,
  samplerType = "RW",
  niter = 10000,
  inits = NULL,
  burn = 0,
  constant_list = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{list, with expected entries "Trait" (corresponding to the trait being modeled by the thermal performance curve)
and "Temp" (corresponding to the temperature in Celcius that the trait is being measured at).}

\item{model}{A string specifying the model name. Use \code{\link[=get_model_names]{get_model_names()}} to view all options.}

\item{priors}{list, optional input specifying prior distributions for parameters (default = NULL).
Elements of the list should correspond to model parameters,
and written using NIMBLE logic. For parameters not specified in the list, default priors are used.}

\item{samplerType}{character string, specifying the sampling method used during the MCMC.
Currently, supported options are:
\itemize{
\item Random Walk Metropolis ('RW')
\item Blocked Random Walk Metropolis ('RW_block')
\item Automated Factor Slice Sampling (AF_slice')
\item Slice sampling ('slice')
}}

\item{niter}{integer, number of MCMC iterations to perform (default is niter = 10000).}

\item{inits}{optional list, initial parameter values to be provided to nimble MCMC.}

\item{burn}{optional integer, number of initial MCMC iterations to be discarded as burn-in. Default is burn = 0}

\item{constant_list}{optional list, constants to be provided to model. If constants are needed and not provided, constant values are used.
Currently only used for model = 'pawar-shsch'.}

\item{verbose}{logical, determines whether to print additional information, Default = TRUE.}

\item{...}{Additional parameters to be passed to nimble during MCMC configuration and sampling.}
}
\value{
\code{b_TPC} returns a list containing entries:
\itemize{
\item \code{samples} - \code{mcmc.list} containing posterior samples of parameters for corresponding model
\item \code{mcmc} -  \code{nimbleModel} object corresponding to model being fit
\item \code{data} -  \code{list} containing trait and temperature data and number of observations (N)
\item \code{model_type} -  \code{character} containing the type of thermal performance curve being fit.
\item \code{constants} - A named vector containing the constant values used, if the model includes constants. Otherwise, returns NULL.
\item \code{uncomp_model} - Uncompiled version of the NIMBLE model. For internal use.
}
}
\description{
Generates NIMBLE model, and performs MCMC.
}
\details{
placeholder
}
\examples{
#placeholder
}
