% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_plotting.R
\name{bayesTPC_ipairs}
\alias{bayesTPC_ipairs}
\title{Wrapper for IDPmisc::ipairs()}
\usage{
bayesTPC_ipairs(
  x,
  burn = 0,
  thin = 1,
  ztransf = function(x) {
     x[x < 1] <- 1
     log2(x)
 },
  ...
)
}
\arguments{
\item{burn}{Integer, number of samples to discard as burn-in before creating pairs plot (default = 0)}

\item{thin}{Integer, thinning interval used to generate pairs plots (default = 1)}

\item{ztransf}{Function, used to transform the counts per pixel in IDPmisc::ipairs.}

\item{...}{additional graphical parameters to be passed as arguments to IDPmisc::ipairs(). For additional information, try ?IDPmisc::ipairs}

\item{object}{Either an object of class `mcmc` or `matrix` OR an object of class `list` which contains an element called `samples` that is of class `mcmc` or `matrix`}
}
\value{
Returns invisible(NULL) and creates pairs plots for MCMC parameters
}
\description{
Wrapper for IDPmisc() that can both directly accept samples of object type `mcmc` or `matrix` as well as an object of class `list` with an element `samples` that is of class `mcmc` or `matrix`
}
\details{
This is a wrapper to create trace plots using coda's `traceplot` function.
}
\examples{
## need data to set up example here. placeholder for now
## set params and reference temperature set
}
