% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_spec.R, R/model_spec.R
\name{change_priors.btpc_likelihood}
\alias{change_priors.btpc_likelihood}
\alias{change_priors}
\alias{change_priors.btpc_model}
\title{Change priors of pre-specified model}
\usage{
\method{change_priors}{btpc_likelihood}(x, priors = character())

change_priors(x, priors)

\method{change_priors}{btpc_model}(x, priors = character())
}
\arguments{
\item{x}{\code{btpc_model} or \code{btpc_likelihood}, The specification to be changed.}

\item{priors}{named character, The names should correspond to the parameters to change, and the values should be the new desired priors.}
}
\value{
Returns the modified model/likelihood. Does not change the default values of any registered model/likelihood type.
}
\description{
Intended to be used to change the priors of an already existing model.
}
\details{
\code{bayesTPC} does not verify if priors specified are compatible with NIMBLE's dialect of BUGS.
All available distributions and formatting are provided on the
\href{https://r-nimble.org/html_manual/cha-writing-models.html#subsec:dists-and-functions}{NIMBLE user manual}.
}
