% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_exception_handling.R
\name{checkParams}
\alias{checkParams}
\title{Checks Prior data format}
\usage{
checkParams(model, params)
}
\arguments{
\item{model}{character, the model type}

\item{params}{named vector, contains the actual params from the function call}
}
\value{
list, input params coerced as a vector, if necessary
}
\description{
Check if data is correctly formatted
}
\details{
TODO
}
