% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUGS_building.R
\name{configure_model}
\alias{configure_model}
\title{Create model string}
\usage{
configure_model(model, priors = NULL, constants = NULL, verbose = T)
}
\arguments{
\item{model}{character or \code{btpc_model}. If a character, a string specifying the model name. Otherwise, a model specification.
If a string is provided, the default values are used if the model is implemented. Use \code{\link[=get_models]{get_models()}} to view all options.}

\item{priors}{list or character, optional input specifying prior distributions for parameters (default = NULL).
Elements of the list should correspond to model parameters, and written using nimble logic.
For parameters not specified in the list, default priors are used.}

\item{constants}{list or character, optional input specifying model constants.
If model requires constant values and none are provided, default values are used.}

\item{verbose}{optional logical. If verbose = TRUE, messages are printed when user-end priors are used, rather than the default values. Default is TRUE.}
}
\value{
character, character string specifying the default model formulation to be passed to \code{nimble}.
}
\description{
Create model string for thermal performance curve model to be passed to nimble.
}
\details{
This function returns a character string of the full \code{nimble} model for a user-specified thermal performance curve and prior distributions.
}
\examples{
## Print default model for briere
cat(configure_model(model = "briere"))

## Use custom prior for 'q' parameter in quadratic curve
my_prior <- list(q = "q~beta(.5, .5)")
cat(configure_model(model = "quadratic", priors = my_prior))
}
