% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_get_functions.R
\name{get_default_constants}
\alias{get_default_constants}
\title{Get Default Model Constants}
\usage{
get_default_constants(model)
}
\arguments{
\item{model}{A string specifying the model name. Use \code{\link[=get_models]{get_models()}} to view all options.}
}
\description{
TODO
}
