% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_get_functions.R
\name{get_default_priors}
\alias{get_default_priors}
\title{Get Default Prior Distributions}
\usage{
get_default_priors(model)
}
\arguments{
\item{model}{A string specifying the model name, or a btpc_model object.
If a string is provided, the default model formula is provided if the model is implemented. If the model is not implemented, an error occurs.
Use \code{\link[=get_models]{get_models()}} to view all options.}
}
\value{
A named list of all parameters and their default prior distribution.
}
\description{
Returns the default distributions for all parameters trained by \code{\link[=b_TPC]{b_TPC()}}.
These are used if no priors are provided.
}
\details{
Prior distributions are written in the format required by NIMBLE's BUGS implementation.
All available distributions and formatting are provided on the
\href{https://r-nimble.org/html_manual/cha-writing-models.html#subsec:dists-and-functions}{NIMBLE user manual}.
}
\examples{
get_default_priors("briere")
}
