% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_formula}
\alias{get_formula}
\alias{get_formulas}
\title{Get Model Formula}
\usage{
get_formula(model)

get_formulas()
}
\arguments{
\item{model}{character or \code{btpc_model}. Use \code{\link[=get_models]{get_models()}} to view all options.}
}
\value{
\code{get_formula()} returns the formula for the provided model as an expression.

\code{get_formulas()} returns a named list containing all models and their respective formulas as expressions.
}
\description{
Returns the formula for a specified model.
}
\details{
\code{get_formula()} is best used to investigate specifics of a model before using it.
It can be used to evaluate models, but it is better practice to use the output of \code{\link[=get_model_function]{get_model_function()}} for direct model evaluation.
}
\examples{
get_formula("stinner")

# If used to evaluate, variables must be set beforehand.
q <- .75
T_max <- 35
T_min <- 10
Temp <- c(15, 20, 25, 30)
ans <- eval(get_formula("quadratic"))
ans

}
\seealso{
\code{\link[=get_model_function]{get_model_function()}}
}
