% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpc_get_functions.R
\name{get_model_constants}
\alias{get_model_constants}
\title{Get Model Constants}
\usage{
get_model_constants(model)
}
\arguments{
\item{model}{A string specifying the model name. Use \code{\link[=get_models]{get_models()}} to view all options.}
}
\value{
If the specified model has constants, a character vector containing all required constants for the model provided.
Otherwise, only outputs a message and returns nothing.
}
\description{
Returns the constants of a specified model, if they exist.
}
\details{
Using this function is only necessary if the model being fit has constant values.
Currently, the only implemented model with constants is 'pawar-shsch'.
}
\examples{

# Model without constants
get_model_constants("gaussian")

# Model with constants
get_model_constants("pawar_shsch")
}
