% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_model_constants}
\alias{get_model_constants}
\alias{get_default_constants}
\title{Get Model Constants}
\usage{
get_model_constants(model)

get_default_constants(model)
}
\arguments{
\item{model}{character or \code{btpc_model}. If a character, a string specifying the model name. Otherwise, a model specification.
If a string is provided, the default values are used if the model is implemented. Use \code{\link[=get_models]{get_models()}} to view all options.}
}
\value{
If the specified model does not contain constants, both \code{get_model_constants()} and \code{get_default_constants()} only output a message and return nothing.

Otherwise, \code{get_model_constants()} returns a character vector containing all required constants for the model provided,
and \code{get_default_constants()}returns a named numeric vector of all default constant values used if none are provided.
}
\description{
Returns the constants of a specified model or their default values, if they exist.
}
\details{
Using this function is only necessary if the model being fit has constant values.
Currently, the only implemented model with constants is 'pawar-shsch'.
}
\examples{

# Model without constants
get_model_constants("gaussian")
get_default_constants("gaussian")

# Model with constants
get_model_constants("pawar_shsch")
get_default_constants("pawar_shsch")
}
