% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_model_function}
\alias{get_model_function}
\title{Get Model as a Function}
\usage{
get_model_function(model, type = "response")
}
\arguments{
\item{model}{character or \code{btpc_model}. Use \code{\link[=get_models]{get_models()}} to view all options.}

\item{type}{character, should the function calculate the link or the response?
Supported inputs are "response" and "link". Default is "response".}
}
\value{
\code{get_model_function} returns a function that evaluates the implemented formula of the specified model.
The parameters of the specified model along with constants (if present) are used as the arguments to the returned function.
}
\description{
Builds and provides a function that evaluates an implemented model.
}
\details{
This function provides a more accessible way to directly evaluate implemented models.
Manual evaluation is available through \code{\link[=get_formula]{get_formula()}}.
}
\examples{

stinner_function <- get_model_function("stinner")
stinner_function

# compare to example in get_formula()
quadratic_function <- get_model_function("quadratic")
quadratic_function(q = .75, T_max = 35, T_min = 10, Temp = c(15, 20, 25, 30))
}
\seealso{
\code{\link[=get_formula]{get_formula()}}
}
