% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_models}
\alias{get_models}
\title{Get Model Names}
\usage{
get_models()
}
\value{
\code{get_models()} returns a character vector of all implemented models in \code{bayesTPC}.
}
\description{
Returns the names of all implemented models in \code{bayesTPC}.
}
