% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_generics.R
\name{plot.btpc_MCMC}
\alias{plot.btpc_MCMC}
\title{Plot an model fit by \code{b_TPC()}.}
\usage{
\method{plot}{btpc_MCMC}(
  x,
  temp_interval = NULL,
  summaryType = "hdi",
  centralSummary = "median",
  prob = 0.9,
  quantiles = c(0.05, 0.95),
  burn = 0,
  type = "response",
  ylab = "Trait",
  legend = TRUE,
  legend_position = "bottomright",
  ...
)
}
\arguments{
\item{x}{\code{btpc_MCMC}, object output from performing MCMC using the \code{bTPC} function.}

\item{temp_interval}{numeric vector, reference values to use to compute values of the thermal performance curve.
If no vector is provided, temp_interval is set as a sequence from the lowest observed temperature in the data to the highest observed temperature in the data, with 1,000 equally spaced points.}

\item{summaryType}{character, type of summary used. Currently supported options are "quantile" and "hdi" (default).}

\item{centralSummary}{character, central summary measure used. Currently supported options are "median" (default) and "mean".}

\item{prob}{numeric, the credible mass used to compute the highest density interval. Used if summaryType = "hdi".}

\item{quantiles}{length 2 numeric, quantiles used for a credible interval. Used if summaryType = "quantile".}

\item{burn}{numeric, initial number of iterations to be discarded as burn-in. Default is 0.}

\item{type}{character, should the summaries be calculated for the link or the response?
Supported inputs are "response" and "link". Default is "response".}

\item{ylab}{a title for the y axis.}

\item{legend}{logical, should a legend be added to the plot? Default is TRUE.}

\item{legend_position}{character, position of the legend. Only used if legend = TRUE. Default is "bottomright".}

\item{...}{additional parameters passed to \code{plot.default()}.}
}
\description{
Plots thermal performance curve summaries using output from \code{\link[=b_TPC]{b_TPC()}}'s nimble MCMC.
}
