% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ppo_plot}
\alias{ppo_plot}
\title{Prior-Posterior Overlap Plot}
\usage{
ppo_plot(
  model,
  burn = 0,
  seq.length = 100,
  legend = TRUE,
  legend_position = "topleft"
)
}
\arguments{
\item{model}{list, usually output from \code{b_TPC()}. \code{ppo_plot} expects this list to have entries "samples", of class \code{mcmc} or \code{numeric} and "priors", with class \code{list} and entries that match the corresponding model parameters.}

\item{burn}{integer, number of samples to discard as burn-in before creating prior-posterior overlap plot. Default is 0.}

\item{seq.length}{integer, length of sequence used to evaluate prior density. Default is 100.}

\item{legend}{logical, should a legend be included? Default is TRUE.}

\item{legend_position}{character, position of the legend. Only used if legend = TRUE. Default is "bottomright".}
}
\value{
Returns invisible(NULL) and creates a prior posterior overlap plot, with the prior density shown using a red line and the posterior density shown using a blue dashed line.
}
\description{
Create a prior-posterior overlap plot from output of \code{\link[=b_TPC]{b_TPC()}}.
}
\details{
Creates a prior-posterior overlap plot. This is useful in determining the impact that the choice of prior distribution has on the analyses.

Priors are evaluated using the \code{setInits()} and \code{calculate()} methods of \code{model$uncomp_model()}. Posteriors are approximated using \code{stats::density()} on the posterior sample of each parameter.
}
