% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_generics.R
\name{summary.btpc_MCMC}
\alias{summary.btpc_MCMC}
\title{Summary statistics for MCMCs}
\usage{
\method{summary}{btpc_MCMC}(object, quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975), ...)
}
\arguments{
\item{object}{\code{btpc_MCMC}, object output from performing MCMC using the \code{bTPC} function.}

\item{quantiles}{numeric, calculated quantiles for each variable.}

\item{...}{additional parameters, passed to \link{summary.mcmc}}
}
\value{
A list (invisible) containing the information about and summary statistics of the MCMC chain(s).
}
\description{
Provide thermal performance curve summaries using output from \code{\link[=b_TPC]{b_TPC()}}'s nimble MCMC.
}
\details{
This function returns various summaries of the output of the thermal performance curve model, generated using MCMC samples from the object returned by \code{b_TPC()}.
}
