% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{traceplot}
\alias{traceplot}
\title{Wrapper for coda::traceplot()}
\usage{
traceplot(object, burn = 0, thin = 1, ...)
}
\arguments{
\item{object}{\code{mcmc}, \code{mcmc.list}, or \code{list}. If a list, must contain an element called \code{samples} that is of class \code{mcmc} or \code{mcmc.list}.}

\item{burn}{integer, number of samples to discard as burn-in before creating traceplot. Default is 0.}

\item{thin}{integer, thinning interval used to generate traceplots.  Default is 1.}

\item{...}{additional graphical parameters to be passed as arguments to coda::traceplot.}
}
\value{
Returns invisible(NULL) and creates trace plots for MCMC parameters.
}
\description{
Wrapper for coda::traceplot() that can both directly accept samples of object type \code{mcmc} or \code{mcmc.list} as well as an object of class \code{list} with an element \code{samples} that is of class \code{mcmc} or \code{mcmc.list}.
}
\details{
This is a wrapper to create trace plots using coda's \code{traceplot()} function.
}
